fix:change some value's type to int in bean/Department.java

+++ src/cn/edu/ncu/newmedia/bean/Department.java
@@ -10,37 +10,37 @@ public class Department {
      * createAt         部门创建时间
      */
 
-    private long id;
-    private long communityId;
+    private int id;
+    private int communityId;
     private String departmentName;
     private Date createAt;
 
     public Department() {
     }
 
-    public Department(long communityId, String departmentName, Date createAt) {
+    public Department(int communityId, String departmentName, Date createAt) {
         this.communityId = communityId;
         this.departmentName = departmentName;
         this.createAt = createAt;
     }
 
-    public Department(long id, long communityId, String departmentName, Date createAt) {
+    public Department(int id, int communityId, String departmentName, Date createAt) {
         this.id = id;
         this.communityId = communityId;
         this.departmentName = departmentName;
         this.createAt = createAt;
     }
 
-    public long getId() {
+    public int getId() {
         return this.id;
     }
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
-    public long getCommunityId() {
+    public int getCommunityId() {
         return this.communityId;
     }
-    public void setCommunityId(long communityId) {
+    public void setCommunityId(int communityId) {
         this.communityId = communityId;
     }
     public String getDepartmentName() {
